可以插入了, 但是自动派单bug, 只能派一个人

+++ infintech-executors/executor-collection-dispatcher/src/main/java/com/infincash/cron/collection/table/TBizCollection.java
@@ -44,7 +44,7 @@ public class TBizCollection {
 
     private String updateBy;
 
-    private Byte state;
+    private int state;
 
     private Date fullRepayDate;
 
@@ -182,11 +182,11 @@ public class TBizCollection {
         this.updateBy = updateBy == null ? null : updateBy.trim();
     }
 
-    public Byte getState() {
+    public int getState() {
         return state;
     }
 
-    public void setState(Byte state) {
+    public void setState(int state) {
         this.state = state;
     }
